test: building 업로드 테스트 (#236)
## 🔥 이슈

- #98 

## 📜 개요

- upload 테스트

## 🤔 PR 유형

어떤 변경 사항이 있나요?

- [ ] 새로운 기능 추가
- [ ] 버그 수정
- [ ] CSS 등 사용자 UI 디자인 변경
- [ ] 코드에 영향을 주지 않는 변경사항(오타 수정, 탭 사이즈 변경, 변수명 변경)
- [ ] 코드 리팩토링
- [ ] 주석 추가 및 수정
- [ ] 문서 수정
- [x] 테스트 추가, 테스트 리팩토링
- [ ] 빌드 부분 혹은 패키지 매니저 수정
- [ ] 파일 혹은 폴더명 수정
- [ ] 파일 혹은 폴더 삭제

## ✅ PR Checklist

PR이 다음 요구 사항을 충족하는지 확인하세요.

- [x] 커밋 메시지 컨벤션에 맞게 작성했습니다.
- [x] 변경 사항에 대한 테스트를 했습니다.(버그 수정/기능에 대한 테스트).

+++ src/main/java/bjs/zangbu/complexList/mapper/ComplexListMapper.java
@@ -4,19 +4,18 @@ import bjs.zangbu.complexList.vo.ComplexList;
 import org.apache.ibatis.annotations.Mapper;
 
 /**
- * 복합 단지(ComplexList) 관련 DB 매핑 인터페이스 (MyBatis Mapper)
- * 단지 정보의 삽입 기능을 담당합니다.
+ * 복합 단지(ComplexList) 관련 DB 매핑 인터페이스 (MyBatis Mapper) 단지 정보의 삽입 기능을 담당합니다.
  */
 @Mapper
 public interface ComplexListMapper {
 
-    /**
-     * 단지(ComplexList) 정보를 데이터베이스에 삽입하고, 생성된 단지 ID를 반환합니다.
-     *
-     * @param complexList 삽입할 단지 정보 객체
-     * @return 삽입 후 생성된 단지 ID
-     */
-    Long createComplexList(ComplexList complexList);
+  /**
+   * 단지(ComplexList) 정보를 데이터베이스에 삽입하고, 생성된 단지 ID를 반환합니다.
+   *
+   * @param complexList 삽입할 단지 정보 객체
+   * @return 삽입 후 생성된 단지 ID
+   */
+  int createComplexList(ComplexList complexList);
 
-    Long  getComplexNoByBuildingId(long buildingId);
+  Long getComplexNoByBuildingId(long buildingId);
 }
